AKHQ-67 Create error page backend - Added ApiError - Created new ErrorController
