+++ src/main/java/org/kafkahq/rest/error/ApiError.java
@@ -1,0 +1,13 @@
+package org.kafkahq.rest.error;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class ApiError {
+    private String title;
+    private String description;
+}
